requirements.txt
mss>=9.0.0
Pillow>=10.0.0
pynput>=1.7.6
PyQt5>=5.15.0
